requirements.txt
PyYAML
aiohttp
aiosqlite
audio_purity_scorer
denoiser
librosa
mutagen
numpy
openpyxl
pandas
pyannote
pydub
requests
resemblyzer
soundfile
speaker_diarization
srt
torch
tqdm
webrtcvad
whisper
yt_uae
